Split query handler into domain-specific modules
- query.ts: 437 → 16 lines, thin registration layer
- query-costs.ts: costs, daily-costs, entity-detail handlers
- query-trends.ts: trend analysis handler
- query-recommendations.ts: savings + missing-tags handlers
- query-filters.ts: filter value enumeration handler
- Move resolveAvailablePeriods to query-utils.ts as shared helper
