+++ packages/desktop/src/main/handlers/query.ts
@@ -1,437 +1,16 @@
 import { ipcMain } from 'electron';
-import {
-  buildCostQuery,
-  buildDailyCostsQuery,
-  buildTrendQuery,
-  buildMissingTagsQuery,
-  buildNonResourceCostQuery,
-  buildEntityDetailQuery,
-  buildSource,
-  buildAliasSqlCase,
-  buildRuleMatchExpr,
-  computePeriodsInRange,
-  listLocalMonths,
-  logger,
-  asEntityRef,
-  asDollars,
-  asDateString,
-  tagColumnName,
-} from '@costgoblin/core';
-import type {
-  CostQueryParams,
-  CostResult,
-  DailyCostsParams,
-  DailyCostsResult,
-  DailyCostDay,
-  Dollars,
-  TrendQueryParams,
-  TrendResult,
-  MissingTagsParams,
-  MissingTagsResult,
-  EntityDetailParams,
-  EntityDetailResult,
-  SavingsResult,
-} from '@costgoblin/core';
 import type { AppContext } from './context.js';
-import type { RawRow } from '../duckdb-client.js';
-import {
-  applyOrgTreeRollup,
-  buildAccountReverseMap,
-  buildCostResult,
-  buildEntityDetailResult,
-  buildMissingTagsResult,
-  buildTrendResult,
-  isOwnerGroupBy,
-  mergeCostRowsByEntity,
-  mergeTrendRowsByEntity,
-  resolveEntityName,
-  toEffort,
-  toNum,
-  toStr,
-} from './query-utils.js';
+import { registerCostHandlers } from './query-costs.js';
+import { registerTrendHandlers } from './query-trends.js';
+import { registerRecommendationHandlers } from './query-recommendations.js';
+import { registerFilterHandlers } from './query-filters.js';
 
 export function registerQueryHandlers(app: AppContext): void {
-  const { ctx, getQueryDimensions: getDimensions, getAccountMap, getOrgAccountsPath, getOrgTreeConfig, getConfig, getCostScope, getAvailableColumns, runQuery, runPreparedQuery } = app;
-
   ipcMain.handle('query:cancel-pending', () => {
-    ctx.db.cancelPendingQueries();
+    app.ctx.db.cancelPendingQueries();
   });
-
-  async function resolveAvailablePeriods(
-    tier: 'daily' | 'hourly',
-    dateRange: { readonly start: string; readonly end: string },
-  ): Promise<{ available: string[]; empty: boolean }> {
-    const available = await listLocalMonths(ctx.dataDir, tier);
-    const required = computePeriodsInRange(dateRange);
-    const usePeriods = required.filter(p => available.includes(p));
-    if (usePeriods.length === 0) {
-      logger.debug('query:plan', { tier, mode: 'empty', requestedMonths: required.length, availableMonths: available.length });
-      return { available, empty: true };
-    }
-    return { available, empty: false };
-  }
-
-  ipcMain.handle('query:costs', async (_event, params: CostQueryParams): Promise<CostResult> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    const orgPath = await getOrgAccountsPath();
-    const costScope = await getCostScope().catch(() => undefined);
-    const tier = params.granularity === 'hourly' ? 'hourly' : 'daily';
-    const availableColumns = await getAvailableColumns(tier);
-    const { available, empty } = await resolveAvailablePeriods(tier, params.dateRange);
-    if (empty) return { rows: [], totalCost: asDollars(0), topServices: [], dateRange: params.dateRange };
-    const { sql, params: queryParams } = buildCostQuery(params, ctx.dataDir, dimensions, undefined, orgPath, available, accountReverseMap, costScope, availableColumns);
-    logger.info('query:costs', { groupBy: params.groupBy });
-
-    const rows = await runPreparedQuery(sql, queryParams);
-    let result = buildCostResult(rows, params.dateRange);
-
-    if (params.groupBy === 'account' || params.groupBy === 'account_id') {
-      // Resolve raw account_id → display name, then merge any rows whose
-      // resolved name collides (strip patterns / normalize on the Account dim
-      // can intentionally collapse N ids into one logical entity).
-      result = {
-        ...result,
-        rows: mergeCostRowsByEntity(result.rows.map(r => ({ ...r, entity: asEntityRef(resolveEntityName(r.entity, accountMap)) }))),
-      };
-    }
-
-    if (isOwnerGroupBy(params.groupBy, dimensions) && params.orgNodeValues === undefined) {
-      const orgTreeConfig = await getOrgTreeConfig();
-      if (orgTreeConfig.tree.length > 0) {
-        result = applyOrgTreeRollup(result, orgTreeConfig.tree);
-      }
-    }
-
-    return result;
-  });
-
-  ipcMain.handle('query:daily-costs', async (_event, params: DailyCostsParams): Promise<DailyCostsResult> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    const orgPath = await getOrgAccountsPath();
-    const costScope = await getCostScope().catch(() => undefined);
-    const tier = params.granularity === 'hourly' ? 'hourly' : 'daily';
-    const availableColumns = await getAvailableColumns(tier);
-    const { available, empty } = await resolveAvailablePeriods(tier, params.dateRange);
-    if (empty) return { days: [], groups: [], totalCost: asDollars(0) };
-    const { sql, params: queryParams } = buildDailyCostsQuery(params, ctx.dataDir, dimensions, orgPath, available, accountReverseMap, costScope, availableColumns);
-    logger.info('query:daily-costs', { groupBy: params.groupBy });
-
-    const rows = await runPreparedQuery(sql, queryParams);
-
-    const dayMap = new Map<string, Record<string, number>>();
-    const groupSet = new Set<string>();
-    let totalCost = 0;
-
-    for (const row of rows) {
-      const rawDate = row['date'];
-      const rawGroup = row['group_name'];
-      let date: string;
-      if (rawDate instanceof Date) {
-        date = rawDate.toISOString().slice(0, 10);
-      } else if (typeof rawDate === 'string') {
-        date = rawDate;
-      } else {
-        date = '';
-      }
-      const group = typeof rawGroup === 'string' ? rawGroup : '';
-      const cost = Number(row['cost'] ?? 0);
-
-      groupSet.add(group);
-      totalCost += cost;
-
-      const existing = dayMap.get(date);
-      if (existing !== undefined) {
-        existing[group] = (existing[group] ?? 0) + cost;
-      } else {
-        dayMap.set(date, { [group]: cost });
-      }
-    }
-
-    const days: DailyCostDay[] = [...dayMap.entries()]
-      .sort((a, b) => a[0].localeCompare(b[0]))
-      .map(([date, breakdown]) => {
-        const total = Object.values(breakdown).reduce((s, v) => s + v, 0);
-        const typedBreakdown: Record<string, Dollars> = {};
-        for (const [k, v] of Object.entries(breakdown)) {
-          typedBreakdown[k] = asDollars(v);
-        }
-        return { date: asDateString(date), total: asDollars(total), breakdown: typedBreakdown };
-      });
-
-    return { days, groups: [...groupSet], totalCost: asDollars(totalCost) };
-  });
-
-  ipcMain.handle('query:trends', async (_event, params: TrendQueryParams): Promise<TrendResult> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    const orgPath = await getOrgAccountsPath();
-    const costScope = await getCostScope().catch(() => undefined);
-    const availableColumns = await getAvailableColumns('daily');
-    const { available, empty } = await resolveAvailablePeriods('daily', params.dateRange);
-    if (empty) return { increases: [], savings: [], totalIncrease: asDollars(0), totalSavings: asDollars(0) };
-    const { sql, params: queryParams } = buildTrendQuery(params, ctx.dataDir, dimensions, orgPath, available, accountReverseMap, costScope, availableColumns);
-    logger.info('query:trends', { groupBy: params.groupBy });
-
-    const rows = await runPreparedQuery(sql, queryParams);
-    const result = buildTrendResult(rows, params.deltaThreshold, params.percentThreshold);
-    if (params.groupBy === 'account' || params.groupBy === 'account_id') {
-      return {
-        ...result,
-        increases: mergeTrendRowsByEntity(result.increases.map(r => ({ ...r, entity: asEntityRef(resolveEntityName(r.entity, accountMap)) }))),
-        savings: mergeTrendRowsByEntity(result.savings.map(r => ({ ...r, entity: asEntityRef(resolveEntityName(r.entity, accountMap)) }))),
-      };
-    }
-    return result;
-  });
-
-  ipcMain.handle('query:missing-tags', async (_event, params: MissingTagsParams): Promise<MissingTagsResult> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    const orgPath = await getOrgAccountsPath();
-    const costScope = await getCostScope().catch(() => undefined);
-    const availableColumns = await getAvailableColumns('daily');
-    logger.info('query:missing-tags', { tagDimension: params.tagDimension });
-
-    const { available, empty } = await resolveAvailablePeriods('daily', params.dateRange);
-    if (empty) {
-      return {
-        rows: [],
-        totalActionableCost: asDollars(0),
-        totalLikelyUntaggableCost: asDollars(0),
-        totalNonResourceCost: asDollars(0),
-        actionableCount: 0,
-        likelyUntaggableCount: 0,
-        nonResourceRows: [],
-      };
-    }
-    const resourceQuery = buildMissingTagsQuery(params, ctx.dataDir, dimensions, orgPath, available, accountReverseMap, costScope, availableColumns);
-    const nonResourceQuery = buildNonResourceCostQuery(params, ctx.dataDir, dimensions, orgPath, available, accountReverseMap, costScope, availableColumns);
-    const [resourceRows, nonResourceRows] = await Promise.all([
-      runPreparedQuery(resourceQuery.sql, resourceQuery.params),
-      runPreparedQuery(nonResourceQuery.sql, nonResourceQuery.params),
-    ]);
-    const result = buildMissingTagsResult(resourceRows, nonResourceRows);
-    return {
-      ...result,
-      rows: result.rows.map(r => ({
-        ...r,
-        accountName: resolveEntityName(r.accountId, accountMap) || r.accountName,
-      })),
-    };
-  });
-
-  ipcMain.handle('query:savings', async (): Promise<SavingsResult> => {
-    const config = await getConfig();
-    const provider = config.providers[0];
-    if (provider?.sync.costOptimization === undefined) {
-      return { recommendations: [], totalMonthlySavings: asDollars(0) };
-    }
-
-    let rows: RawRow[];
-    try {
-      rows = await runQuery(`
-        SELECT
-          account_id,
-          account_name,
-          action_type,
-          current_resource_type,
-          COALESCE(recommended_resource_summary, '') AS summary,
-          COALESCE(region, '') AS region,
-          COALESCE(estimated_monthly_savings_after_discount, 0) AS monthly_savings,
-          COALESCE(estimated_monthly_cost_after_discount, 0) AS monthly_cost,
-          COALESCE(estimated_savings_percentage_after_discount, 0) AS savings_pct,
-          COALESCE(implementation_effort, '') AS effort,
-          COALESCE(resource_arn, '') AS resource_arn,
-          COALESCE(current_resource_details, '') AS current_details,
-          COALESCE(recommended_resource_details, '') AS recommended_details,
-          COALESCE(current_resource_summary, '') AS current_summary,
-          COALESCE(restart_needed, false) AS restart_needed,
-          COALESCE(rollback_possible, false) AS rollback_possible,
-          COALESCE(recommendation_source, '') AS recommendation_source
-        FROM read_parquet('${ctx.dataDir}/aws/raw/cost-opt-*/*.parquet', filename=true)
-        QUALIFY ROW_NUMBER() OVER (PARTITION BY recommendation_id ORDER BY filename DESC) = 1
-        ORDER BY monthly_savings DESC
-      `);
-    } catch {
-      return { recommendations: [], totalMonthlySavings: asDollars(0) };
-    }
-
-    let totalSavings = 0;
-    const recommendations = rows.map(r => {
-      const savings = toNum(r['monthly_savings']);
-      totalSavings += savings;
-      return {
-        accountId: toStr(r['account_id']),
-        accountName: toStr(r['account_name']),
-        actionType: toStr(r['action_type']),
-        resourceType: toStr(r['current_resource_type']),
-        summary: toStr(r['summary']),
-        region: toStr(r['region']),
-        monthlySavings: asDollars(savings),
-        monthlyCost: asDollars(toNum(r['monthly_cost'])),
-        savingsPercentage: toNum(r['savings_pct']),
-        effort: toEffort(toStr(r['effort'])),
-        resourceArn: toStr(r['resource_arn']),
-        currentDetails: toStr(r['current_details']),
-        recommendedDetails: toStr(r['recommended_details']),
-        currentSummary: toStr(r['current_summary']),
-        restartNeeded: Boolean(r['restart_needed']),
-        rollbackPossible: Boolean(r['rollback_possible']),
-        recommendationSource: toStr(r['recommendation_source']),
-      };
-    });
-
-    return { recommendations, totalMonthlySavings: asDollars(totalSavings) };
-  });
-
-  ipcMain.handle('query:entity-detail', async (_event, params: EntityDetailParams): Promise<EntityDetailResult> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    const orgPath = await getOrgAccountsPath();
-    const costScope = await getCostScope().catch(() => undefined);
-    const tier = params.granularity === 'hourly' ? 'hourly' : 'daily';
-    const availableColumns = await getAvailableColumns(tier);
-    const { available, empty } = await resolveAvailablePeriods(tier, params.dateRange);
-    if (empty) {
-      return {
-        entity: params.entity,
-        totalCost: asDollars(0),
-        previousCost: asDollars(0),
-        percentChange: 0,
-        dailyCosts: [],
-        byAccount: [],
-        byService: [],
-        bySubEntity: [],
-      };
-    }
-    const { sql, params: queryParams } = buildEntityDetailQuery(params, ctx.dataDir, dimensions, orgPath, available, accountReverseMap, costScope, availableColumns);
-    logger.info('query:entity-detail', { entity: params.entity });
-
-    const rows = await runPreparedQuery(sql, queryParams);
-    const result = buildEntityDetailResult(rows, params.entity);
-    return {
-      ...result,
-      byAccount: result.byAccount.map(s => ({
-        ...s,
-        name: resolveEntityName(s.name, accountMap),
-      })),
-    };
-  });
-
-  ipcMain.handle('query:filter-values', async (_event, dimensionId: string, filterEntries: Record<string, string>, dateRange?: { start: string; end: string }, opts?: { bypassCostScope?: boolean }): Promise<{ value: string; label: string; count: number }[]> => {
-    const dimensions = await getDimensions();
-    const accountMap = await getAccountMap();
-    const accountReverseMap = buildAccountReverseMap(accountMap);
-    // Honour the cost-scope exclusions here too — a user who has
-    // excluded Tax / RI purchases doesn't want those values showing up
-    // in filter dropdowns either. The Cost Scope editor sets
-    // bypassCostScope=true so its rule-condition autocomplete still
-    // sees every available value when composing new rules.
-    const costScope = opts?.bypassCostScope === true
-      ? undefined
-      : await getCostScope().catch(() => undefined);
-
-    const builtIn = dimensions.builtIn.find(d => d.name === dimensionId);
-    const tag = dimensions.tags.find(d => tagColumnName(d.tagName) === dimensionId);
-
-    const field = builtIn === undefined ? dimensionId : builtIn.field;
-    // Apply alias/normalize CASE for both built-ins and tags so the SQL
-    // returns the same display values the cost queries see — including the
-    // SSM-derived friendly region names spliced into Region's aliases.
-    let fieldExpr = field;
-    if (builtIn !== undefined) {
-      fieldExpr = buildAliasSqlCase(field, builtIn);
-    } else if (tag !== undefined) {
-      fieldExpr = buildAliasSqlCase(field, tag);
-    }
-
-    const whereClauses: string[] = [];
-    for (const [key, value] of Object.entries(filterEntries)) {
-      const fb = dimensions.builtIn.find(d => d.name === key);
-      const ft = dimensions.tags.find(d => tagColumnName(d.tagName) === key);
-      const ff = fb === undefined ? key : fb.field;
-      let ffExpr = ff;
-      if (fb !== undefined) {
-        ffExpr = buildAliasSqlCase(ff, fb);
-      } else if (ft !== undefined) {
-        ffExpr = buildAliasSqlCase(ff, ft);
-      }
-      // Account filters carry display names that may collapse N ids — same
-      // expansion the SQL builder does in buildFilterClauses.
-      if (ff === 'account_id') {
-        const ids = accountReverseMap.get(value);
-        if (ids !== undefined && ids.length > 0) {
-          const list = ids.map(id => `'${id.replaceAll("'", "''")}'`).join(', ');
-          whereClauses.push(`${ff} IN (${list})`);
-          continue;
-        }
-      }
-      whereClauses.push(`${ffExpr} = '${value.replaceAll("'", "''")}'`);
-    }
-
-    if (dateRange !== undefined) {
-      whereClauses.push(`usage_date BETWEEN '${dateRange.start}' AND '${dateRange.end}'`);
-    }
-
-    if (costScope !== undefined) {
-      // NOT (rule_match) per enabled rule — same shape buildExclusionClauses
-      // emits for the main query builders. Apply the dim's reverse map
-      // (account collapse) when the condition targets account_id.
-      for (const rule of costScope.rules) {
-        if (!rule.enabled) continue;
-        const matchExpr = buildRuleMatchExpr(rule, dimensions, accountReverseMap);
-        if (matchExpr === null) continue;
-        whereClauses.push(`NOT (${matchExpr})`);
-      }
-    }
-
-    const whereStr = whereClauses.length > 0 ? `WHERE ${whereClauses.join(' AND ')}` : '';
-
-    const orgPath = await getOrgAccountsPath();
-    const periods = dateRange === undefined ? undefined : computePeriodsInRange(dateRange);
-    const availableColumns = await getAvailableColumns('daily');
-    // filter-values doesn't need a specific metric — just needs source
-    // with the fewest column constraints. Pass 'unblended' (always
-    // present) plus the probed column set so the source query doesn't
-    // reference missing columns.
-    const source = buildSource(ctx.dataDir, 'daily', dimensions, orgPath, periods, 'unblended', availableColumns);
-    const sql = `
-      SELECT ${fieldExpr} AS val, SUM(cost) AS total_cost
-      FROM ${source}
-      ${whereStr}
-      GROUP BY val
-      HAVING val IS NOT NULL AND val != ''
-      ORDER BY total_cost DESC
-      LIMIT 100
-    `;
-
-    const rows = await runQuery(sql);
-    const isAccountDim = dimensionId === 'account' || dimensionId === 'account_id';
-    if (isAccountDim) {
-      // Roll N raw ids that resolve to the same display name into one entry —
-      // value AND label are the display name so the picker shows a single row
-      // and downstream filter matching uses the same display name (which the
-      // SQL builder expands back to all underlying ids).
-      const merged = new Map<string, number>();
-      for (const r of rows) {
-        const rawVal = toStr(r['val']);
-        const name = accountMap.get(rawVal) ?? rawVal;
-        merged.set(name, (merged.get(name) ?? 0) + toNum(r['total_cost']));
-      }
-      return [...merged.entries()]
-        .map(([name, cost]) => ({ value: name, label: name, count: cost }))
-        .sort((a, b) => b.count - a.count);
-    }
-    return rows.map(r => {
-      const rawVal = toStr(r['val']);
-      return { value: rawVal, label: rawVal, count: toNum(r['total_cost']) };
-    });
-  });
+  registerCostHandlers(app);
+  registerTrendHandlers(app);
+  registerRecommendationHandlers(app);
+  registerFilterHandlers(app);
 }
